Split VMR tests
